# Pool more JIT resources to reduce memory usage/contention