--- a/doc/src/devdocs/locks.md
+++ b/doc/src/devdocs/locks.md
@@ -28,7 +28,6 @@ The following are definitely leaf locks (level 1), and must not try to acquire a
 >   * gc_perm_lock
 >   * flisp
 >   * jl_in_stackwalk (Win32)
->   * PM_mutex[i]
 >   * ResourcePool<?>::mutex
 >
 >     > flisp itself is already threadsafe, this lock only protects the `jl_ast_context_list_t` pool